refactor: utilsへ移動
