+++ src/app/ar_assets/create/_components/StepContent/CompletedArAsset.tsx
@@ -3,9 +3,9 @@ import { Container } from '@/shared/components/common/Container';
 import { Notification } from '@/shared/components/common/Feedback';
 import { Center, Stack } from '@/shared/components/common/Layout';
 import { QRCode } from '@/shared/components/common/QRCode';
-import { rem } from '@/shared/components/common/Style';
 import { Text } from '@/shared/components/common/Text';
 import { IconArrowBarToDown, IconCheck } from '@/shared/components/icons';
+import { rem } from '@/shared/util/converter';
 
 export const CompletedArAsset = () => {
   const checkIcon = <IconCheck style={{ width: rem(20), height: rem(20) }} />;
