--- src/shared/components/common/Style/rem.tsx
@@ -1,3 +1,0 @@
-import { rem as Mantinerem } from '@mantine/core';
-
-export const rem = Mantinerem;
