+++ src/shared/utils/converter/rem.ts
@@ -1,0 +1,3 @@
+import { rem as remConverter } from '@mantine/core';
+
+export const rem = remConverter;
